+++ my-app/src/components/Pages/UserGalleryPage.js
@@ -3,16 +3,7 @@ import { useState, useEffect } from 'react';
 import { useParams, Link } from 'react-router-dom';
 import { Text, Flex, Center } from '@chakra-ui/react';
 import FeedGrid from '../shared/Feed/FeedGrid';
-
-const PaginationNav = ({ pageNumber, prevURL, nextURL }) => (
-  <Center mt={2} mb={2} fontSize="lg">
-    <Link to={prevURL}>◀️</Link>
-    <Text mr={1} ml={1}>
-      {pageNumber}
-    </Text>
-    <Link to={nextURL}>▶️</Link>
-  </Center>
-);
+import PaginationNav from '../shared/Feed/PaginationNav';
 
 export default function UserGalleryPage() {
   const [data, setData] = useState([]);
